ml/src/document_classification/requirements.txt
# Document Classification Requirements
# Python 3.9+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
efficientnet-pytorch>=0.7.1

# Image Processing
Pillow>=9.5.0
opencv-python>=4.8.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.2.0

# Model Explainability
shap>=0.42.0

# Visualization & Monitoring
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database (PostgreSQL)
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
